requirements.txt
openai-whisper
yt-dlp
browser-cookie3
torch
